realized my carousel componenet was not responsive with my previous css
